Split assembly part appender resolution from SqlStatementBuilder into a specific part appender factory
