+++ src/HatTrick.DbEx.Sql/Assembler/DelegateSqlStatementBuilderFactory.cs
@@ -1,20 +1,17 @@
 ﻿using HatTrick.DbEx.Sql.Configuration;
 using HatTrick.DbEx.Sql.Expression;
-using HatTrick.DbEx.Sql.Pipeline;
 using System;
-using System.Collections.Generic;
-using System.Text;
 
 namespace HatTrick.DbEx.Sql.Assembler
 {
     public class DelegateSqlStatementBuilderFactory : ISqlStatementBuilderFactory
     {
         #region internals
-        private readonly Func<SqlStatementAssemblerConfiguration, QueryExpression, IAppender, ISqlParameterBuilder, ISqlStatementBuilder> factory;
+        private readonly Func<IAssemblyPartAppenderFactory, SqlStatementAssemblerConfiguration, QueryExpression, IAppender, ISqlParameterBuilder, ISqlStatementBuilder> factory;
         #endregion
 
         #region constructors
-        public DelegateSqlStatementBuilderFactory(Func<SqlStatementAssemblerConfiguration, QueryExpression, IAppender, ISqlParameterBuilder, ISqlStatementBuilder> factory)
+        public DelegateSqlStatementBuilderFactory(Func<IAssemblyPartAppenderFactory, SqlStatementAssemblerConfiguration, QueryExpression, IAppender, ISqlParameterBuilder, ISqlStatementBuilder> factory)
         {
             this.factory = factory ?? throw new DbExpressionConfigurationException($"{nameof(factory)} is required to initialize a Sql Statement Builder."); ;
         }
@@ -24,9 +21,9 @@ namespace HatTrick.DbEx.Sql.Assembler
             if (factory is null)
                 throw new DbExpressionConfigurationException($"{nameof(factory)} is required to initialize a Sql Statement Builder.");
 
-            this.factory = new Func<SqlStatementAssemblerConfiguration, QueryExpression, IAppender, ISqlParameterBuilder, ISqlStatementBuilder>((c, s, a, p) =>
+            this.factory = new Func<IAssemblyPartAppenderFactory, SqlStatementAssemblerConfiguration, QueryExpression, IAppender, ISqlParameterBuilder, ISqlStatementBuilder>((partAppenderFactory, config, expression, appender, parameterBuilder) =>
             {
-                var f = factory().CreateSqlStatementBuilder(c, s, a, p);
+                var f = factory().CreateSqlStatementBuilder(partAppenderFactory, config, expression, appender, parameterBuilder);
                 if (f is null)
                     throw new DbExpressionException("Cannot create a Sql Statement Builder: The factory returned a null value.");
                 return f;
@@ -35,8 +32,8 @@ namespace HatTrick.DbEx.Sql.Assembler
         #endregion
 
         #region methods
-        public ISqlStatementBuilder CreateSqlStatementBuilder(SqlStatementAssemblerConfiguration config, QueryExpression expression, IAppender appender, ISqlParameterBuilder parameterBuilder)
-            => factory(config, expression, appender, parameterBuilder);
+        public ISqlStatementBuilder CreateSqlStatementBuilder(IAssemblyPartAppenderFactory partAppenderFactory, SqlStatementAssemblerConfiguration config, QueryExpression expression, IAppender appender, ISqlParameterBuilder parameterBuilder)
+            => factory(partAppenderFactory, config, expression, appender, parameterBuilder);
         #endregion
     }
 }
