+++ src/HatTrick.DbEx.Sql/Assembler/IAssemblyPartAppenderFactory.cs
@@ -1,0 +1,7 @@
+﻿namespace HatTrick.DbEx.Sql.Assembler
+{
+    public interface IAssemblyPartAppenderFactory
+    {
+        IAssemblyPartAppender CreatePartAppender(object part);
+    }
+}
