+++ src/HatTrick.DbEx.Sql/Assembler/ISqlStatementBuilderFactory.cs
@@ -1,16 +1,14 @@
-﻿using HatTrick.DbEx.Sql.Builder;
-using HatTrick.DbEx.Sql.Configuration;
-using HatTrick.DbEx.Sql.Executor;
+﻿using HatTrick.DbEx.Sql.Configuration;
 using HatTrick.DbEx.Sql.Expression;
-using System;
 
 namespace HatTrick.DbEx.Sql.Assembler
 {
     public interface ISqlStatementBuilderFactory
     {
         ISqlStatementBuilder CreateSqlStatementBuilder(
+            IAssemblyPartAppenderFactory partAppenderFactory,
             SqlStatementAssemblerConfiguration config,
-            QueryExpression expression, 
+            QueryExpression expression,
             IAppender appender,
             ISqlParameterBuilder parameterBuilder
         );
