+++ src/HatTrick.DbEx.Sql/Expression/ExpressionContainer.cs
@@ -3,6 +3,7 @@
 namespace HatTrick.DbEx.Sql.Expression
 {
     public class ExpressionContainer :
+        IExpression,
         IEquatable<ExpressionContainer>
     {
         #region interface
